requirements.txt
# Django
django>=2.1.5
django-environ


# Models
django-polymorphic


# Admin
django-nested-admin

# Caching
django-redis


# Event Queueing, Web workers
gunicorn

# Databases
psycopg2-binary
djongo

# Cloud
cloudinary
boto3
django-storages


# Image Support
Pillow


# WYSIWYG Editor
django-ckeditor


# Authentication + Permissions
django-role-permissions
django-registration
django-otp
django-two-factor-auth
twilio
qrcode


# REST API Support
coreapi
django-rest-swagger
djangorestframework>=3.7.7
djangorestframework-httpsignature
djangorestframework-jsonapi>=2.6.0
djangorestframework_simplejwt
django-filter>=2.0.0
django-guardian
django-oauth-plus
django-oauth2-provider
oauth2


# Server
django-livereload-server
django-cors-headers
django-debug-toolbar


# Chats
channels

# Seeding data
Faker
pycountry

# Text analysis
textdistance

# Markdown Support
markdown
